Dynamically generate a group of tabs for an rmd chunk

Chunk options must contain results = 'asis'. Function writes directly to the chunk using side-effects

Usage

bespoke_rmd_tab_group(
  df,
  inputspec,
  outputspec,
  alert_rules = NULL,
  timepoint_limits = c(NA, NA),
  fill_with_zero = FALSE,
  tab_order = NULL,
  tab_group_name = NULL,
  tab_group_level = 1
)

Arguments

df

Data frame containing time series in long format

inputspec

inputspec() object specifying which columns in the supplied df represent the "timepoint", "item", "value" and "tab" for the time series. A separate tab will be created for each distinct value in the "tab" column.

outputspec

outputspec object specifying the desired format of the html table(s). If not supplied, default values will be used.

alert_rules

alert_rules() object specifying conditions to test

timepoint_limits

Set start and end dates for time period to include. Defaults to min/max of timepoint_col

fill_with_zero

Logical. Replace any missing or NA values with 0? Useful when value_col is a record count

tab_order

Optional vector containing values from tab_col in desired order of display

tab_group_name

Character string to appear on parent tab label. If omitted or NULL, only the child tabs (and not the parent tab) will be created.

tab_group_level

Integer specifying the nesting level of the parent tab. Value of 1 equates to rmd level "##". This is unaffected by the presence or not of tab_group_name.

Value

(invisibly) the supplied df

Details

The supplied data frame should contain multiple time series in long format, i.e.:

one "timepoint" (date) column which will be used for the x-axes. This currently must be at a daily granularity, but values do not have to be consecutive.

one "item" (character) column containing categorical values identifying distinct time series.

one "value" (numeric) column containing the time series values which will be used for the y-axes.

Optionally, a "tab" (character) column containing categorical values which will be used to group the time series into different tabs on the report.

The inputspec parameter maps the data frame columns to the above.