Example data frame containing multiple time series in long format

Simulated data to cover a range of different behaviours of time series

Usage

example_data

Format

example_data

A data frame with 3,903 rows and 4 columns:

timepoint - Dates for the time series

item - Labels to identify the different time series

value - Values for the time series

tab - Labels to group related time series into tabs